# bookstore-front/src/app/service/model/book.ts
/**
 * BookStore APIs
 * BookStore APIs exposed from a Java EE back-end to an Angular front-end
 *
 * OpenAPI spec version: 1.0.0
 * Contact: abdeldjallil.maiza@gmail.com
 *
 * NOTE: This class is auto generated by the swagger code generator program.
 * https://github.com/swagger-api/swagger-codegen.git
 * Do not edit the class manually.
 */


/**
 * Book resource representation
 */
export interface Book {
    /**
     * Identifier
     */
    id?: number;
    /**
     * ISBN number
     */
    isbn: string;
    /**
     * Title of the book
     */
    title: string;
    /**
     * URL of the image cover
     */
    imgUrl?: string;
    /**
     * Language in which the book has been written
     */
    language?: Book.LanguageEnum;
    /**
     * Unit cost
     */
    unitCost?: number;
    /**
     * Number of pages
     */
    nbOfPages?: number;
    publicationDate?: Date;
    /**
     * Summary describing the book
     */
    description?: string;
}
export namespace Book {
    export type LanguageEnum = 'ENGLISH' | 'FRENCH' | 'SPANISH' | 'PORTUGUESE' | 'ITALIAN' | 'FINISH' | 'GERMAN' | 'DEUTSCH' | 'RUSSIAN';
    export const LanguageEnum = {
        ENGLISH: 'ENGLISH' as LanguageEnum,
        FRENCH: 'FRENCH' as LanguageEnum,
        SPANISH: 'SPANISH' as LanguageEnum,
        PORTUGUESE: 'PORTUGUESE' as LanguageEnum,
        ITALIAN: 'ITALIAN' as LanguageEnum,
        FINISH: 'FINISH' as LanguageEnum,
        GERMAN: 'GERMAN' as LanguageEnum,
        DEUTSCH: 'DEUTSCH' as LanguageEnum,
        RUSSIAN: 'RUSSIAN' as LanguageEnum
    }
}